requirements.txt
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
autopep8==2.0.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.4.1
exceptiongroup==1.1.2
fastapi==0.100.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
kombu==5.3.1
numpy==1.25.2
prompt-toolkit==3.0.39
pycodestyle==2.11.0
pydantic==2.1.1
pydantic_core==2.4.0
pymongo==4.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.1
vine==5.0.0
wcwidth==0.2.6
